pyproject.toml
[project]
name = "quant-qmt-proxy"
version = "0.1.0"
description = ""
authors = [{ name = "zhzhl0", email = "zhzhl0@163.com" }]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "grpcio (>=1.76.0,<2.0.0)",
    "grpcio-tools (>=1.76.0,<2.0.0)",
    "protobuf (>=6.33.0,<7.0.0)",
    "grpcio-reflection (>=1.76.0,<2.0.0)",
    "xtquant (>=250516.1.1,<250517.0.0)",
    "pytest-asyncio (>=0.23.0,<0.24.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
priority = "primary"

# poetry add fastapi uvicorn[standard] pydantic pydantic-settings python-multipart python-dotenv httpx loguru PyYAML grpcio grpcio-tools protobuf grpcio-reflection xtquant
